+++ src/card_collections/mod.rs
@@ -4,6 +4,6 @@ mod foundations;
 mod freecell;
 
 pub use self::card_collection::CardCollection;
-pub use self::cascade::{Cascade, Cascades};
+pub use self::cascade::{parse_cascade, Cascade, Cascades};
 pub use self::foundations::{Foundation, Foundations};
 pub use self::freecell::{parse_freecells, Freecell, Freecells};